// lib/types/database.types.ts
export type Json = string | number | boolean | null | { [key: string]: Json | undefined } | Json[]

export interface Database {
  public: {
    Tables: {
      users: {
        Row: {
          id: string
          email: string
          full_name: string
          role: string
          created_at: string
          updated_at: string | null
        }
        Insert: {
          id: string
          email: string
          full_name: string
          role: string
          created_at?: string
          updated_at?: string | null
        }
        Update: {
          id?: string
          email?: string
          full_name?: string
          role?: string
          created_at?: string
          updated_at?: string | null
        }
      }
      lawyer_profiles: {
        Row: {
          id: string
          user_id: string
          specialization: string | null
          experience: number | null
          hourly_rate: number | null
          bio: string | null
          created_at: string
          updated_at: string | null
          availability: Json | null
        }
        Insert: {
          id?: string
          user_id: string
          specialization?: string | null
          experience?: number | null
          hourly_rate?: number | null
          bio?: string | null
          created_at?: string
          updated_at?: string | null
          availability?: Json | null
        }
        Update: {
          id?: string
          user_id?: string
          specialization?: string | null
          experience?: number | null
          hourly_rate?: number | null
          bio?: string | null
          created_at?: string
          updated_at?: string | null
          availability?: Json | null
        }
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
  }
}